+++ src/pages/Mypage/admin/PostCRUD/PostCRUDPage.module.css
@@ -1,43 +1,43 @@
-.post-upload-section-form {
+.post-section-form {
   display: flex;
   flex-direction: column;
   gap: 2.5rem;
 }
-.post-upload-section-form .title {
+.post-section-form .title {
   margin-bottom: 1rem;
   font-size: 1.2rem;
 }
 
-.post-upload-section-form h3.title {
+.post-section-form h3.title {
   margin-top: 2.5rem;
 }
 
-.post-upload-section-form > section > fieldset {
+.post-section-form > section > fieldset {
   margin-top: 1.5rem;
 }
 
-.upload-imgs-wrapper {
+.update-imgs-wrapper {
   display: flex;
   gap: 1rem;
 }
-.upload-img-input {
+.update-img-input {
   cursor: pointer;
   flex-shrink: 0;
   background-color: var(--gray200-color);
   border-radius: 0.5rem;
   margin-bottom: 1rem;
 }
-.upload-img-input,
+.update-img-input,
 .img-cover {
   width: 15vw;
   min-width: 130px;
   max-width: 10rem;
   aspect-ratio: 3/4;
 }
-.upload-img-input input {
+.update-img-input input {
   display: none;
 }
-.upload-img-input svg {
+.update-img-input svg {
   position: relative;
   top: 50%;
   left: 50%;
@@ -103,4 +103,22 @@
 
 .tags-section :global(.tags-input) {
   width: 100%;
+}
+.tags-section :global(.tagify__tag) {
+  max-width: calc(100% - 10px);
+}
+.tags-section :global(.tagify__tag > div::before) {
+  --tag-inset-shadow-size: 10rem;
+}
+.tags-section :global(.tagify__tag-text) {
+  word-break: break-all;
+}
+.tags-section :global(.tagify__tag__removeBtn) {
+  flex-shrink: 0;
+}
+
+.button-group {
+  display: flex;
+  gap: 1rem;
+  justify-content: space-between;
 }
