--- src/pages/Mypage/admin/PostUpdate/PostUpdatePage.module.css
@@ -1,124 +1,0 @@
-.post-update-section-form {
-  display: flex;
-  flex-direction: column;
-  gap: 2.5rem;
-}
-.post-update-section-form .title {
-  margin-bottom: 1rem;
-  font-size: 1.2rem;
-}
-
-.post-update-section-form h3.title {
-  margin-top: 2.5rem;
-}
-
-.post-update-section-form > section > fieldset {
-  margin-top: 1.5rem;
-}
-
-.update-imgs-wrapper {
-  display: flex;
-  gap: 1rem;
-}
-.update-img-input {
-  cursor: pointer;
-  flex-shrink: 0;
-  background-color: var(--gray200-color);
-  border-radius: 0.5rem;
-  margin-bottom: 1rem;
-}
-.update-img-input,
-.img-cover {
-  width: 15vw;
-  min-width: 130px;
-  max-width: 10rem;
-  aspect-ratio: 3/4;
-}
-.update-img-input input {
-  display: none;
-}
-.update-img-input svg {
-  position: relative;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-.imgs-list {
-  overflow-x: scroll;
-  display: flex;
-}
-.imgs-list li {
-  margin-right: 0.5rem;
-}
-
-.img-cover {
-  border: 1px solid var(--gray200-color);
-  border-radius: 0.5rem;
-  overflow: hidden;
-  position: relative;
-}
-.img-cover img {
-  width: 100%;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  object-fit: contain;
-}
-
-.l_date {
-  display: grid;
-  grid-template-areas: "a b c";
-  gap: 1rem;
-}
-@media (max-width: 420px) {
-  .l_date {
-    flex-direction: column;
-    grid-template-areas:
-      "a b"
-      "c c";
-  }
-}
-
-.l_address {
-  display: flex;
-  gap: 1rem;
-}
-
-@media (max-width: 440px) {
-  .l_address {
-    flex-wrap: wrap;
-  }
-}
-
-.l_address fieldset {
-  flex-grow: 4;
-}
-
-.l_address button {
-  flex-grow: 1;
-  flex-basis: 100px;
-  flex-shrink: 0;
-}
-
-.tags-section :global(.tags-input) {
-  width: 100%;
-}
-.tags-section :global(.tagify__tag) {
-  max-width: calc(100% - 10px);
-}
-.tags-section :global(.tagify__tag > div::before) {
-  --tag-inset-shadow-size: 10rem;
-}
-.tags-section :global(.tagify__tag-text) {
-  word-break: break-all;
-}
-.tags-section :global(.tagify__tag__removeBtn) {
-  flex-shrink: 0;
-}
-
-.button-group {
-  display: flex;
-  gap: 1rem;
-  justify-content: space-between;
-}
