+++ src/pages/index.tsx
@@ -10,8 +10,7 @@ import PayFailPage from "./PayFail/PayFailPage";
 import Mypage from "./Mypage/Mypage";
 import ProfilePage from "./Mypage/ProfilePage/ProfilePage";
 import PostManagePage from "./Mypage/admin/PostManage/PostManagePage";
-import PostUpdatePage from "./Mypage/admin/PostUpdate/PostUpdatePage";
-import PostUploadPage from "./Mypage/admin/PostUpload/PostUploadPage";
+import PostCRUDPage from "./Mypage/admin/PostCRUD/PostCRUDPage";
 import AdminReservationManagePage from "./Mypage/admin/ReservationManage/ReservationManagePage";
 import LikeManagePage from "./Mypage/user/LikeManage/LikeManagePage";
 import ReviewManagePage from "./Mypage/user/ReviewManage/ReviewManagePage";
@@ -31,8 +30,7 @@ export {
   Mypage,
   ProfilePage,
   PostManagePage,
-  PostUpdatePage,
-  PostUploadPage,
+  PostCRUDPage,
   AdminReservationManagePage,
   UserReservationManagePage,
   LikeManagePage,
